requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
attrs==19.3.0
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==5.4.3
selenium==3.141.0
six==1.15.0
urllib3==1.25.9
wcwidth==0.2.5
